STBT-3643 Toaster UI process hangs Reason for change: none Test Procedure: none Risks: none
Signed-off-by: Eugene Baranov <ebaranov@productengine.com>

===== ORIGINAL MESSAGE =====
process: Deal with infinite looping of the server

Currently if an exception occurs, we just run the idle handler again and
again, usually looping indefintely. Chances are the exception that occurred
will keep occuring and this is not a good place to be.

This was breaking the autobuilders with gigabytes of logs.

At least improve things so the cooker shuts down gracefully when this happens.
Some trace of the original problem may still be present on the console too!

Change-Id: Ie9707dc849f20ea004b510ec596034933409da55
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>