STBT-3643 Toaster UI process hangs Reason for change: none Test Procedure: none Risks: none
Signed-off-by: Eugene Baranov <ebaranov@productengine.com>

===== ORIGINAL MESSAGE =====
process: Further improve robustness against server shutdown

Currently, if an exception occurs in an event handler, the server shuts
down but the UI simply hangs. This happens in two places, firstly waiting
for events and secondly, sending events to a server which no longer exists.

The latter does time out, the former does not. These patches improve
both code sections to check if the main server process is alive and if not,
trigger things to shut down gracefully. This avoids the timeout in the
command sending case too.

This resolves various cases where the UI would simply hang indefintely.

Change-Id: I8c2d1d499235ff88eaa0e77a0a7d852ffc4d03dd
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>